# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "**TMF API Reference : TMF - 724 Incident**\n\n**Release : 21.5 - June\
    \ 2021**\n\nThe Incident API provides a standardized client interface to Incident\
    \ Management Systems for creating, tracking and managing Incidents as a result\
    \ of an issue or problem identified by a customer or another system. \nExamples\
    \ of Incident API originators (clients) include CRM applications, network management\
    \ or fault management systems, or other Incident management systems (e.g. B2B).\n\
    \nThe API supports the ability to send requests to create a new Incident specifying\
    \ the nature and severity of the trouble or issue as well as all necessary related\
    \ information. The API also includes mechanisms to search for and update existing\
    \ Incidents. Notifications are defined to provide information when a Incident\
    \ has been updated, including status changes. A basic set of states of a Incident\
    \ has been specified (as an example) to handle Incident lifecycle management.\n\
    Incidenting API manages Incident resource:\n\n  -\tA Incident represents a record,\
    \ or an issue raised by requestor that need to be solved, used for reporting and\
    \ managing the resolution of problems, incidents or request -\tMain Incident attributes\
    \ are its description, severity, type, related dates (creation, expected resolution,\
    \ resolution), state and related information (change reason and change date),\
...
host: "serverRoot:8080"
basePath: "/tmf-api/Incident/v4/"
...
schemes:
...
consumes:
...
produces:
...
paths:
  /incident:
    get:
      tags:
      - "incident"
      summary: "List or find Incident objects"
      description: "This operation list or find Incident entities"
      operationId: "listIncident"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Incident"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Incident"
    post:
      tags:
      - "incident"
      summary: "Creates a Incident"
      description: "This operation creates a Incident entity."
      operationId: "createIncident"
      parameters:
      - in: "body"
        name: "incident"
        description: "The Incident to be created"
        required: true
        schema:
          $ref: "#/definitions/Incident_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Incident"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Incident"
  /incident/{id}:
    get:
      tags:
      - "incident"
      summary: "Retrieves a Incident by ID"
      description: "This operation retrieves a Incident entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveIncident"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Incident"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Incident"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Incident"
    delete:
      tags:
      - "incident"
      summary: "Deletes a Incident"
      description: "This operation deletes a Incident entity."
      operationId: "deleteIncident"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Incident"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Incident"
    patch:
      tags:
      - "incident"
      summary: "Updates partially a Incident"
      description: "This operation updates partially a Incident entity."
      operationId: "patchIncident"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Incident"
        required: true
        type: "string"
      - in: "body"
        name: "incident"
        description: "The Incident to be updated"
        required: true
        schema:
          $ref: "#/definitions/Incident_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/Incident"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Incident"
...
definitions:
...
  Incident:
    type: "object"
    properties:
      id:
        type: "string"
        description: "unique identifier"
      href:
        type: "string"
        format: "uri"
        description: "Hyperlink reference"
      ackTime:
        type: "string"
        description: "The acktime of the incident, for example 2020-03-10T04:01:12Z"
      category:
        type: "string"
        description: "The category of the incident, according to the define of ITU"
      clearTime:
        type: "string"
        description: "The clear time  of the incident, for example 2020-03-10T04:01:12Z"
      detail:
        type: "string"
        description: "The detail description of the incident"
      domain:
        type: "string"
        description: "The domain of the incident, for example RAN, PON, OTN, Cross-Domain\
          \ etc"
      impact:
        type: "string"
        description: "The impact  of the incident, impact  enumlation：extensive/significant/moderate/minor"
      name:
        type: "string"
        description: "The name of the incident."
      occurTime:
        type: "string"
        description: "The occur time  of the incident, for example 2020-03-10T04:01:12Z"
      priority:
        type: "string"
        description: "The priority  of the incident, priority  critical/high/medium/low"
      repairAdvice:
        type: "string"
        description: "if the incident can be repaid, give the advice"
      status:
        type: "string"
        description: "The status time  of the incident, for example unacknowledged-and-uncleared"
      updateTime:
        type: "string"
        description: "The last update time  of the incident, for example 2020-03-10T04:01:12Z"
      urgency:
        type: "string"
        description: "The urgency  of the incident, urgency  enumlation：critical/high/medium/lowr"
      affects:
        type: "array"
        description: "The  incident affect object, may be one service or customer\
          \ ONT"
        items:
          $ref: "#/definitions/ResourceEntity"
      events:
        type: "array"
        description: "The correlation event object such as alarm, perfomance or something\
          \ else etc"
        items:
          $ref: "#/definitions/Event"
      extensionInfo:
        type: "array"
        description: "The  is use for the extension of incident"
        items:
          $ref: "#/definitions/Characteristic"
      rootCauses:
        type: "array"
        description: "The rootCauses show the incident, it's may be several rootCauses"
        items:
          $ref: "#/definitions/RootCause"
      sourceObjects:
        type: "array"
        description: "The objects show the incident, it's may be part of Network Equipment"
        items:
          $ref: "#/definitions/ResourceEntity"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "An Incident is a record of an event that has altered the operational\
      \ state of a entity (Resource, Service or Customers Product)"
    example:
      ackTime: "ackTime"
      extensionInfo:
      - characteristicRelationship:
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        valueType: "valueType"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
        value: {}
      - characteristicRelationship:
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        valueType: "valueType"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
        value: {}
      clearTime: "clearTime"
      '@type': "@type"
      impact: "impact"
      occurTime: "occurTime"
      affects:
      - '@baseType': "@baseType"
        '@type': "@type"
        id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
        href: "http://myservice.com/resource-path"
        '@schemaLocation': "http://example.com/aeiou"
      - '@baseType': "@baseType"
        '@type': "@type"
        id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
        href: "http://myservice.com/resource-path"
        '@schemaLocation': "http://example.com/aeiou"
      updateTime: "updateTime"
      priority: "priority"
      repairAdvice: "repairAdvice"
      sourceObjects:
      - '@baseType': "@baseType"
        '@type': "@type"
        id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
        href: "http://myservice.com/resource-path"
        '@schemaLocation': "http://example.com/aeiou"
      - '@baseType': "@baseType"
        '@type': "@type"
        id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
        href: "http://myservice.com/resource-path"
        '@schemaLocation': "http://example.com/aeiou"
      rootCauses:
      - rootCauseObject:
          '@baseType': "@baseType"
          '@type': "@type"
          id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
          href: "http://myservice.com/resource-path"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "http://example.com/aeiou"
        detail: "detail"
        '@schemaLocation': "http://example.com/aeiou"
      - rootCauseObject:
          '@baseType': "@baseType"
          '@type': "@type"
          id: "aac9969d-219d-4ff1-b256-1765dcf9b342"
          href: "http://myservice.com/resource-path"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "http://example.com/aeiou"
        detail: "detail"
        '@schemaLocation': "http://example.com/aeiou"
      '@baseType': "@baseType"
      urgency: "urgency"
      domain: "domain"
      name: "name"
      id: "id"
      href: "http://example.com/aeiou"
      detail: "detail"
      category: "category"
      '@schemaLocation': "http://example.com/aeiou"
      events:
      - eventId: "eventId"
        '@type': "@type"
        analyticCharacteristic:
        - characteristicRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
        - characteristicRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
        description: "description"
        eventType: "eventType"
        source:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        priority: "priority"
        title: "title"
        relatedParty:
        - '@referredType': "@referredType"
          role: "role"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        - '@referredType': "@referredType"
          role: "role"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        domain: "domain"
        eventTime: "2000-01-23T04:56:07.000+00:00"
        reportingSystem:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        correlationId: "correlationId"
        id: "id"
        href: "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
        timeOccurred: "2000-01-23T04:56:07.000+00:00"
      - eventId: "eventId"
        '@type': "@type"
        analyticCharacteristic:
        - characteristicRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
        - characteristicRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
        description: "description"
        eventType: "eventType"
        source:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        priority: "priority"
        title: "title"
        relatedParty:
        - '@referredType': "@referredType"
          role: "role"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        - '@referredType': "@referredType"
          role: "role"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        '@baseType': "@baseType"
        domain: "domain"
        eventTime: "2000-01-23T04:56:07.000+00:00"
        reportingSystem:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
        correlationId: "correlationId"
        id: "id"
        href: "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
        timeOccurred: "2000-01-23T04:56:07.000+00:00"
      status: "status"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "string"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "string"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"